# docs/swagger.yaml
basePath: /
definitions:
  serializer.CartJSON:
    properties:
      consumption_id:
        type: integer
      modes_count:
        type: integer
    type: object
...
  serializer.FuelConsumptionDetailJSON:
    properties:
      consumption_id:
        type: integer
      creator_login:
        type: string
      date_completed:
        type: string
      date_create:
        type: string
      date_formed:
        type: string
      destination:
        type: string
      entries:
        items:
          $ref: '#/definitions/serializer.FuelConsumptionEntryViewJSON'
        type: array
      fuel_price:
        type: number
      moderator_login:
        type: string
      origin:
        type: string
      status:
        type: string
      total_saved:
        description: ✅ Остался float64
        type: number
    type: object
  serializer.FuelConsumptionEntryViewJSON:
    properties:
      base_consumption:
        type: number
      economy_percent:
        type: number
      fuel_saved:
        type: number
      id:
        type: integer
      image_key:
        type: string
      mode_id:
        type: integer
      mode_name:
        type: string
      route_distance:
        type: integer
    type: object
  serializer.FuelConsumptionListJSON:
    properties:
      consumption_id:
        type: integer
      creator_login:
        type: string
      date_completed:
        type: string
      date_create:
        type: string
      date_formed:
        type: string
      destination:
        type: string
      fuel_entries_count:
        type: integer
      fuel_price:
        type: number
      moderator_login:
        type: string
      origin:
        type: string
      status:
        type: string
      total_saved:
        description: ✅ Остался float64 (не *float64)
        type: number
    type: object
...
  serializer.FuelConsumptionUpdateJSON:
    properties:
      destination:
        type: string
      fuel_price:
        type: number
      origin:
        type: string
    type: object
  serializer.ModeJSON:
    properties:
      base_consumption:
        type: number
      description:
        type: string
      driving_type:
        type: string
      economy_percent:
        type: number
      image_key:
        type: string
      is_active:
        type: boolean
      mode_id:
        type: integer
      mode_name:
        type: string
      video_key:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчёта экономии топлива
  title: Cruise control API
  version: "1.0"
paths:
  /api/Modes:
    post:
      consumes:
      - multipart/form-data
      description: multipart/form-data или application/json; только для модераторов
      parameters:
      - description: Название режим
        in: formData
        name: Mode_title
        type: string
      - description: Коэффициент материала
        in: formData
        name: Mode_material_coefficient
        type: number
      - description: Коэффициент толщины
        in: formData
        name: Mode_thickness_coefficient
        type: number
      - description: Описание
        in: formData
        name: description
        type: string
      - description: Фото режима
        in: formData
        name: photo
        type: file
      - description: Видео режима
        in: formData
        name: video
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/serializer.ModeJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуется роль модератора
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать режим
      tags:
      - Modes
...
  /api/fuel-consumptions:
    get:
      description: Фильтры по дате и статусу. Пользователь видит свои; модератор —
        все.
      parameters:
      - description: Начало диапазона даты (YYYY-MM-DD)
        in: query
        name: created_at
        type: string
      - description: Конец диапазона даты (YYYY-MM-DD)
        in: query
        name: date_completed
        type: string
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.FuelConsumptionListJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Список заявок на проверку экономии
      tags:
      - fuel_consumptions
  /api/fuel_consumptions/{id}:
    delete:
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку
      tags:
      - fuel_consumptions
    get:
      description: 'Данные по заявке; доступ: создатель или модератор.'
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.FuelConsumptionDetailJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Заявка по ID
      tags:
      - fuel_consumptions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.FuelConsumptionUpdateJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.FuelConsumptionListJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить заявку
      tags:
      - fuel_consumptions
...
  /api/fuel_consumptions/{id}/form:
    put:
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.FuelConsumptionListJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Оформить заявку
      tags:
      - fuel_consumptions
  /api/fuel_consumptions/cart:
    get:
      description: Если черновика нет — возвращает статус "no_draft".
      produces:
      - application/json
      responses:
        "200":
          description: CartJSON (tire_pressure_id, tires_count) или no_draft
          schema:
            $ref: '#/definitions/serializer.CartJSON'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Корзина (черновик заявки)
      tags:
      - fuel_consumptions
  /api/modes:
    get:
      parameters:
      - description: Поиск по названию режима
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.ModeJSON'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список режимов
      tags:
      - Modes
...
  /fuel_consumption/{id}:
    get:
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML-страница заявки
          schema:
            type: string
        "302":
          description: Редирект при ошибке или отсутствии доступа
      summary: Страница заявки на проверку экономии
      tags:
      - fuel_consumption-html
...
  /modes:
    get:
      parameters:
      - description: Поиск по названию режима
        in: query
        name: query
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML-страница
          schema:
            type: string
      summary: Главная страница (типы движения)
      tags:
      - mode-html
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"